requirements.txt
requests==2.31.0
# 센서 라이브러리 (실 기기 사용 시)
python-telegram-bot==13.15
RPi.GPIO==0.7.1   # (라즈베리파이 OS 전용, Windows에서 설치 어려울 수 있음)
Adafruit_DHT==1.4.0
smbus2==0.4.3


# Raspberry Pi에서 GPIO를 제어하려면 **RPi.GPIO** 라이브러리가 필요합니다.  
# **Adafruit_DHT**(혹은 `adafruit-circuitpython-dht`) 등도 설치 시, 라즈비안/Ubuntu 환경에서의 종속 라이브러리가 필요할 수 있습니다(예: `sudo apt-get install libgpiod2` 등).
